+++ packages/gov-ui-kit/src/modules/components/member/memberDataListItem/memberDataListItemSkeleton/memberDataListItemSkeleton.stories.tsx
@@ -3,7 +3,7 @@ import { MemberDataListItem } from '../../../..';
 import { DataList } from '../../../../../core';
 
 const meta: Meta<typeof MemberDataListItem.Skeleton> = {
-    title: 'Modules/Components/Member/MemberDataListItem.Skeleton',
+    title: 'Modules/Components/Member/MemberDataListItem/MemberDataListItem.Skeleton',
     component: MemberDataListItem.Skeleton,
     tags: ['autodocs'],
     parameters: {