+++ packages/gov-ui-kit/src/modules/components/transaction/transactionDataListItem/transactionDataListItemSkeleton/index.ts
@@ -1,0 +1,4 @@
+export {
+    TransactionDataListItemSkeleton,
+    type ITransactionDataListItemSkeletonProps,
+} from './transactionDataListItemSkeleton';